Track list length as we go.
Counting the length of the list was a whopping 22% of total time spent in the
benchmark. However, removing this doesn't save us 22%: fold_left is
substantially more expensive than partition_tf, because map is optimized under
the hood. We however still save about 10% of total time between tracking the
length and skipping a reverse pass.