Use an array and partition its elements in place.
We can batch allocations by splatting all the strings into an array and doing
the partition_tf work in place on that array. In practice this turns out to be
the fastest approach: we don't have to look at very many elements, so it's
faster than sorting or converting all elements to integers upfront. If the input
became large enough, such that pointer chasing became a substantial part of the
work, an upfront integer conversion might be preferable.